integration_test: code cleanup + speedup + fixes + improvements
* restructure the code, basically putting things into functions. this will make it easier to cache things later
* speedup computation by a) removing the 2 slowests classifiers b) subsampling the features c) subsampling the samples
* fixing some minor bugs (like `numerai_dataset` in one place and `numerai_datasets` in another)
* improving the tests by fetching `originality` and `concordance` with the API instead of checking manually

Obviously, this still doesn't run in a minute. But it should already be much faster + more automated. In addition, the code is in better shaper for future optimizations. I cannot test this end to end, as the testing API is not accessible. This is also using the old API and might need changes..